fix(entire-workspace-dashboard): fix test code according to the scenario
